requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.4.0
numpy==1.26.4
faiss-cpu==1.8.0.post1; python_version<"3.13"
sentence-transformers==2.7.0
langgraph>=0.2.0
openai>=1.43.0
pytest==8.3.3
python-dotenv==1.0.1
scikit-learn>=1.0.0
